Fixed CurrencyAmountView crash
Crash happens because of hacked values of precision constants introduced as a workaround for BTC being introduced into exchange rates, requiring a more precise display compared to the other currencies.

Undid the hack and enforced a check on currency code whenever exchange rate is displayed. If BTC, precision is forced to 8 digits.